refactor(core): Creating pruning service
